requirements.txt
# Copyright (C) 2022 Universidade de Aveiro, DETI/IEETA, Bioinformatics Group - http://bioinformatics.ua.pt/
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==1.4.9
    # via kombu
anyjson==0.3.3
    # via kombu
arabic-reshaper==2.1.3
    # via xhtml2pdf
billiard==3.3.0.23
    # via celery
biopython==1.76
    # via -r requirements.in
celery==3.1.26.post2
    # via
    #   -r requirements.in
    #   django-celery
certifi==2020.4.5.1
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via
    #   django-newsletter
    #   requests
configparser==4.0.2
    # via arabic-reshaper
cryptography==3.3.2
    # via
    #   jwcrypto
    #   pyopenssl
    #   pysaml2
cssselect==1.1.0
    # via premailer
cssutils==1.0.2
    # via premailer
defusedxml==0.7.1
    # via
    #   djangosaml2
    #   pysaml2
deprecated==1.2.13
    # via jwcrypto
django-ace==1.0.11
    # via -r requirements.in
django-adminplus==0.5
    # via -r requirements.in
django-allauth==0.40.0
    # via -r requirements.in
django-appconf==1.0.3
    # via
    #   -r requirements.in
    #   django-compressor
django-bootstrap-breadcrumbs==0.9.2
    # via -r requirements.in
django-bootstrap-pagination==1.7.1
    # via -r requirements.in
django-celery==3.3.1
    # via -r requirements.in
django-compressor==2.4.1
    # via -r requirements.in
django-constance[database]==2.5.0
    # via -r requirements.in
django-contrib-comments==1.9.2
    # via -r requirements.in
django-countries==5.5
    # via -r requirements.in
django-crispy-forms==1.8.1
    # via -r requirements.in
django-extensions==1.9.9
    # via -r requirements.in
django-filter==1.1.0
    # via -r requirements.in
django-grappelli==2.12.4
    # via -r requirements.in
django-guardian==1.5.1
    # via
    #   -r requirements.in
    #   django-userena-ce
git+https://github.com/bioinformatics-ua/django-hitcount.git#egg=django-hitcount
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-markup-deprecated==0.0.4
    # via -r requirements.in
git+https://github.com/bioinformaticsua/django-neue-transmeta#egg=django-neue-transmeta
    # via -r requirements.in
django-newsletter==0.8b1
    # via -r requirements.in
django-picklefield==2.1.1
    # via django-constance
django-premailer==0.2.0
    # via -r requirements.in
django-resized==0.3.11
    # via -r requirements.in
django-sslserver==0.20
    # via -r requirements.in
django-tables2==2.0.6
    # via -r requirements.in
git+https://github.com/bioinformatics-ua/django-userena.git@84b1fa9610860596dd7c95c1f1af24b3942c8ff4#egg=userena
    # via -r requirements.in
django-webpack-loader==0.7.0
    # via -r requirements.in
django==1.11.29
    # via
    #   -r requirements.in
    #   django-ace
    #   django-allauth
    #   django-appconf
    #   django-celery
    #   django-contrib-comments
    #   django-js-reverse
    #   django-newsletter
    #   django-picklefield
    #   django-sslserver
    #   django-tables2
    #   django-userena-ce
    #   djangosaml2
    #   drf-nested-routers
    #   easy-thumbnails
djangorestframework==3.9.4
    # via
    #   -r requirements.in
    #   drf-nested-routers
djangosaml2==0.18.1
    # via -r requirements.in
drf-nested-routers==0.91
    # via -r requirements.in
easy-thumbnails==2.6.0
    # via
    #   -r requirements.in
    #   django-userena-ce
enum34==1.1.10
    # via cryptography
et-xmlfile==1.0.1
    # via openpyxl
furl==2.1.3
    # via recaptcha2
future==0.18.2
    # via arabic-reshaper
geopy==0.99
    # via -r requirements.in
github3.py==1.0.0a4
    # via -r requirements.in
html2text==2019.8.11
    # via django-userena-ce
html5lib==1.1
    # via xhtml2pdf
idna==2.10
    # via requests
ipaddress==1.0.23
    # via cryptography
jdcal==1.4.1
    # via openpyxl
jwcrypto==0.9.1
    # via -r requirements.in
kombu==3.0.37
    # via celery
ldif3==3.1.1
    # via django-newsletter
lxml==4.7.1
    # via premailer
ndg-httpsclient==0.5.1
    # via github3.py
numpy==1.16.6
    # via biopython
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==2.6.4
    # via -r requirements.in
orderedmultidict==1.0.1
    # via furl
pillow==6.2.2
    # via
    #   django-newsletter
    #   easy-thumbnails
    #   reportlab
    #   xhtml2pdf
premailer==3.0.0
    # via django-premailer
psycopg2-binary==2.8.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   github3.py
    #   ndg-httpsclient
pycparser==2.21
    # via cffi
pymongo==2.9.5
    # via -r requirements.in
pyopenssl==21.0.0
    # via
    #   github3.py
    #   ndg-httpsclient
    #   pysaml2
pyparsing==1.5.7
    # via -r requirements.in
pypdf2==1.26.0
    # via xhtml2pdf
pysaml2==4.9.0
    # via
    #   -r requirements.in
    #   djangosaml2
pysolr==3.8.1
    # via -r requirements.in
python-bidi==0.4.2
    # via xhtml2pdf
python-card-me==0.9.3
    # via django-newsletter
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pysaml2
    #   python-card-me
python-levenshtein==0.12.2
    # via -r requirements.in
python-memcached==1.59
    # via -r requirements.in
python-openid==2.2.5
    # via django-allauth
pytz==2021.3
    # via
    #   celery
    #   django
    #   pysaml2
rarfile==2.8
    # via -r requirements.in
rcssmin==1.0.6
    # via django-compressor
recaptcha2==0.1
    # via -r requirements.in
reportlab==3.5.59
    # via xhtml2pdf
requests-oauthlib==1.3.1
    # via django-allauth
requests==2.27.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   github3.py
    #   premailer
    #   pysaml2
    #   pysolr
    #   recaptcha2
    #   requests-oauthlib
rjsmin==1.1.0
    # via django-compressor
selenium==3.141
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
    #   django-appconf
    #   django-bootstrap-breadcrumbs
    #   django-compressor
    #   django-constance
    #   django-contrib-comments
    #   django-countries
    #   django-extensions
    #   django-newsletter
    #   furl
    #   html5lib
    #   jwcrypto
    #   orderedmultidict
    #   pyopenssl
    #   pysaml2
    #   python-bidi
    #   python-dateutil
    #   python-memcached
    #   textile
    #   xhtml2pdf
sorl-thumbnail==12.5.0
    # via django-newsletter
surlex==0.2.0
    # via django-newsletter
textile==2.3.12
    # via -r requirements.in
typing==3.10.0.0
    # via django-extensions
unicodecsv==0.14.1
    # via django-newsletter
uritemplate.py==3.0.2
    # via github3.py
uritemplate==3.0.1
    # via uritemplate.py
urllib3==1.26.8
    # via
    #   requests
    #   selenium
webencodings==0.5.1
    # via html5lib
wrapt==1.13.3
    # via deprecated
wsgiref==0.1.2
    # via -r requirements.in
xhtml2pdf==0.2.5
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
